fix: Resolve test concurrency issue
Parallelism is persisted at the package level. Therefore, the inconsistent behavior was caused by executing tests in parallel (thread concurrency). Since we had different packages (availability, price, and integration), they were recognized by the Go testing module as separate, even though they were pointing to the same test package.
