+++ src/test/integration/find_by_host_integration_test.go
@@ -1,4 +1,4 @@
-package test
+package integration
 
 import (
 	"bookem-room-service/client/userclient"
